@@ -61,16 +61,18 @@ namespace Coding_Conventions_Examples
 
             // Save snippet 4 and 5 for possible additions in program structure.
 
-            Name[] nameList = {new Name { FirstName = "Anderson", LastName = "Redmond" },
-                                 new Name { FirstName = "Jones", LastName = "Seattle" },
-                                  new Name { FirstName = "Anderson", LastName = "Redmond" }};
+            Name[] nameList = [
+                new Name { FirstName = "Anderson", LastName = "Redmond" },
+                new Name { FirstName = "Jones", LastName = "Seattle" },
+                new Name { FirstName = "Anderson", LastName = "Redmond" }
+            ];
             int n = 0;
 
             //<snippet6>
             string displayName = $"{nameList[n].LastName}, {nameList[n].FirstName}";
             //</snippet6>
 
-            Console.WriteLine("{0}, {1}", nameList[n].LastName, nameList[n].FirstName);
+            Console.WriteLine($"{nameList[n].LastName}, {nameList[n].FirstName}");
             Console.WriteLine(nameList[n].LastName + ", " + nameList[n].FirstName);
 
             //<snippet7>
@@ -123,12 +125,9 @@ namespace Coding_Conventions_Examples
             Console.WriteLine();
             //</snippet12>
 
-            //<snippet13a>
-            string[] vowels1 = { "a", "e", "i", "o", "u" };
-            //</snippet13a>
-            //<snippet13b>
-            var vowels2 = new string[] { "a", "e", "i", "o", "u" };
-            //</snippet13b>
+            //<snippet13>
+            string[] vowels = [ "a", "e", "i", "o", "u" ];
+            //</snippet13>
 
             //<snippet15b>
             Del exampleDel2 = DelMethod;
@@ -152,10 +151,7 @@ namespace Coding_Conventions_Examples
             }
             finally
             {
-                if (bodyStyle != null)
-                {
-                    ((IDisposable)bodyStyle).Dispose();
-                }
+                bodyStyle?.Dispose();
             }
             //</snippet17a>
             //<snippet17b>
@@ -214,20 +210,14 @@ namespace Coding_Conventions_Examples
 
             ExampleClass.totalInstances = 1;
 
-            var customers = new List<Customer>
-            {
-              new Customer { Name = "Jones", ID = 432, City = "Redmond" }
-            };
+            List<Customer> Customers = [ new Customer { Name = "Jones", ID = 432, City = "Redmond" } ];
 
             // Check shop name to use this.
-            var distributors = new List<Distributor>
-            {
-              new Distributor { Name = "ShopSmart", ID = 11302, City = "Redmond" }
-            };
+            List<Distributor> Distributors = [ new Distributor { Name = "ShopSmart", ID = 11302, City = "Redmond" } ];
 
             //<snippet25>
             //<snippet28>
-            var seattleCustomers = from customer in customers
+            var seattleCustomers = from customer in Customers
                                    //</snippet28>
                                    where customer.City == "Seattle"
                                    select customer.Name;
@@ -235,20 +225,20 @@ namespace Coding_Conventions_Examples
 
             //<snippet26>
             var localDistributors =
-                from customer in customers
-                join distributor in distributors on customer.City equals distributor.City
+                from customer in Customers
+                join distributor in Distributors on customer.City equals distributor.City
                 select new { Customer = customer, Distributor = distributor };
             //</snippet26>
 
             //<snippet27>
             var localDistributors2 =
-                from customer in customers
-                join distributor in distributors on customer.City equals distributor.City
-                select new { CustomerName = customer.Name, DistributorID = distributor.ID };
+                from customer in Customers
+                join distributor in Distributors on customer.City equals distributor.City
+                select new { CustomerName = customer.Name, DistributorName = distributor.Name };
             //</snippet27>
 
             //<snippet29>
-            var seattleCustomers2 = from customer in customers
+            var seattleCustomers2 = from customer in Customers
                                     where customer.City == "Seattle"
                                     orderby customer.Name
                                     select customer;
@@ -257,13 +247,13 @@ namespace Coding_Conventions_Examples
             // #30 is in class CompoundFrom
 
             var customerDistributorNames =
-                from customer in customers
-                join distributor in distributors on customer.City equals distributor.City
+                from customer in Customers
+                join distributor in Distributors on customer.City equals distributor.City
                 select new { CustomerName = customer.Name, DistributorID = distributor.ID };
 
             var customerDistributorNames2 =
-                from customer in customers
-                from distributor in distributors
+                from customer in Customers
+                from distributor in Distributors
                 where customer.City == distributor.City
                 select new { CustomerName = customer.Name, DistributorID = distributor.ID };
 
@@ -335,20 +325,14 @@ namespace Coding_Conventions_Examples
             return totalInstances;
         }
 
-        public static int ResultSoFar()
-        {
-            return 0;
-        }
+        public static int ResultSoFar() => 0;
     }
 
     class BaseClass
     {
         protected static int totalInstances;
 
-        static BaseClass()
-        {
-            totalInstances = 0;
-        }
+        static BaseClass() => totalInstances = 0;
 
         public static int IncrementTotal()
         {
@@ -367,7 +351,7 @@ namespace Coding_Conventions_Examples
         public class Student
         {
             public string? LastName { get; set; }
-            public List<int>? Scores { get; set; }
+            public ICollection<int> Scores { get; set; } = default!;
         }
 
         static void Main()
@@ -375,32 +359,37 @@ namespace Coding_Conventions_Examples
 
             // Use a collection initializer to create the data source. Note that
             // each element in the list contains an inner sequence of scores.
-            List<Student> students = new List<Student>
-        {
-           new Student {LastName="Omelchenko", Scores= new List<int> {97, 72, 81, 60}},
-           new Student {LastName="O'Donnell", Scores= new List<int> {75, 84, 91, 39}},
-           new Student {LastName="Mortensen", Scores= new List<int> {88, 94, 65, 85}},
-           new Student {LastName="Garcia", Scores= new List<int> {97, 89, 85, 82}},
-           new Student {LastName="Beebe", Scores= new List<int> {35, 72, 91, 70}}
-        };
+            List<Student> students = [
+                new Student {LastName="Omelchenko", Scores = [97, 72, 81, 60]},
+                new Student {LastName="O'Donnell", Scores = [75, 84, 91, 39]},
+                new Student {LastName="Mortensen", Scores = [88, 94, 65, 85]},
+                new Student {LastName="Garcia", Scores = [97, 89, 85, 82]},
+                new Student {LastName="Beebe", Scores = [35, 72, 91, 70]}
+            ];
 
             //<snippet30>
             var scoreQuery = from student in students
-                             from score in student.Scores!
+                             from score in student.Scores
                              where score > 90
                              select new { Last = student.LastName, score };
             //</snippet30>
 
+            // <interpolatedStrings>
             // Execute the queries.
             Console.WriteLine("scoreQuery:");
             foreach (var student in scoreQuery)
             {
-                Console.WriteLine("{0} Score: {1}", student.Last, student.score);
+                Console.WriteLine($"{student.Last} Score: {student.score}");
             }
+            // </interpolatedStrings>
 
-            // Keep the console window open in debug mode.
-            Console.WriteLine("Press any key to exit.");
-            Console.ReadKey();
+            // <rawStringLiterals>
+            var message = """
+                This is a long message that spans across multiple lines.
+                It uses raw string literals. This means we can 
+                also include characters like \n and \t without escaping them.
+                """;
+            // </rawStringLiterals>
         }
     }
 }
@@ -459,7 +448,7 @@ namespace GenericTypeParameters
         //</TypeParametersOne>
 
         //<TypeParametersTwo>
-        public int IComparer<T>() { return 0; }
+        public int IComparer<T>() => 0;
         public delegate bool Predicate<T>(T item);
         public struct Nullable<T> where T : struct { /*...*/ }
         //</TypeParametersTwo>
@@ -474,4 +463,23 @@ namespace GenericTypeParameters
             //</TypeParametersThree>
         }
     }//WrapParameters
+}
+
+namespace Constructors
+{
+    // <PrimaryRecord>
+    public record Person(string FirstName, string LastName);
+    // </PrimaryRecord>
+
+    // <PrimaryClass>
+    public class LabelledContainer<T>(string label)
+    {
+        public string Label { get; } = label;
+        public required T Contents 
+        { 
+            get;
+            init;
+        }
+    }
+    // </PrimaryClass>
 }
