@@ -4,8 +4,10 @@ using Azure.Core.Diagnostics;
 using Azure.Identity;
 using Microsoft.Extensions.Azure;
 
-var userAssignedClientId = "<user-assigned-client-id>";
-var builder = WebApplication.CreateBuilder(args);
+string userAssignedClientId = "<user-assigned-client-id>";
+string storageAccountName = "<account-name>";
+string keyVaultName = "<vault-name>";
+WebApplicationBuilder builder = WebApplication.CreateBuilder(args);
 
 #region snippet_FilteredLogging
 using AzureEventSourceListener listener = new((args, message) =>
@@ -20,8 +22,10 @@ using AzureEventSourceListener listener = new((args, message) =>
 #region snippet_Dac
 builder.Services.AddAzureClients(clientBuilder =>
 {
+    clientBuilder.AddSecretClient(
+        new Uri($"https://{keyVaultName}.vault.azure.net"));
     clientBuilder.AddBlobServiceClient(
-        new Uri("https://<account-name>.blob.core.windows.net"));
+        new Uri($"https://{storageAccountName}.blob.core.windows.net"));
 
     DefaultAzureCredential credential = new();
     clientBuilder.UseCredential(credential);
@@ -31,15 +35,17 @@ builder.Services.AddAzureClients(clientBuilder =>
 #region snippet_DacExcludes
 builder.Services.AddAzureClients(clientBuilder =>
 {
+    clientBuilder.AddSecretClient(
+        new Uri($"https://{keyVaultName}.vault.azure.net"));
     clientBuilder.AddBlobServiceClient(
-        new Uri("https://<account-name>.blob.core.windows.net"));
+        new Uri($"https://{storageAccountName}.blob.core.windows.net"));
 
     clientBuilder.UseCredential(new DefaultAzureCredential(
         new DefaultAzureCredentialOptions
         {
             ExcludeEnvironmentCredential = true,
+            ExcludeManagedIdentityCredential = true,
             ExcludeWorkloadIdentityCredential = true,
-            ManagedIdentityClientId = userAssignedClientId,
         }));
 });
 #endregion snippet_DacExcludes
@@ -47,11 +53,13 @@ builder.Services.AddAzureClients(clientBuilder =>
 #region snippet_Ctc
 builder.Services.AddAzureClients(clientBuilder =>
 {
+    clientBuilder.AddSecretClient(
+        new Uri($"https://{keyVaultName}.vault.azure.net"));
     clientBuilder.AddBlobServiceClient(
-        new Uri("https://<account-name>.blob.core.windows.net"));
+        new Uri($"https://{storageAccountName}.blob.core.windows.net"));
 
     clientBuilder.UseCredential(new ChainedTokenCredential(
-        new ManagedIdentityCredential(clientId: userAssignedClientId),
+        new AzurePowerShellCredential(),
         new VisualStudioCredential()));
 });
 #endregion snippet_Ctc
@@ -67,7 +75,8 @@ TokenCredential credential;
 
 if (app.Environment.IsProduction() || app.Environment.IsStaging())
 {
-    credential = new ManagedIdentityCredential(clientId: userAssignedClientId);
+    credential = new ManagedIdentityCredential(
+        ManagedIdentityId.FromUserAssignedClientId(userAssignedClientId));
 }
 else
 {
@@ -82,17 +91,16 @@ credential = new DefaultAzureCredential(
     {
         ExcludeEnvironmentCredential = true,
         ExcludeWorkloadIdentityCredential = true,
-        ExcludeAzureCliCredential = true,
+        ExcludeManagedIdentityCredential = true,
         ExcludeAzurePowerShellCredential = true,
         ExcludeAzureDeveloperCliCredential = true,
-        ManagedIdentityClientId = userAssignedClientId
     });
 #endregion
 
 #region snippet_CtcEquivalents
 credential = new ChainedTokenCredential(
-    new ManagedIdentityCredential(clientId: userAssignedClientId),
-    new VisualStudioCredential());
+    new VisualStudioCredential(),
+    new AzureCliCredential());
 #endregion
 
 if (app.Environment.IsDevelopment())
