@@ -11,12 +11,11 @@ public static class ReturnQueryFromMethod
             select i.ToString();
 
         void QueryMethod2(int[] ints, out IEnumerable<string> returnQ) =>
-            returnQ =
-                from i in ints
-                where i < 4
-                select i.ToString();
+            returnQ = from i in ints
+                      where i < 4
+                      select i.ToString();
 
-        int[] nums = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
+        int[] nums = [ 0, 1, 2, 3, 4, 5, 6, 7, 8, 9 ];
 
         var myQuery1 = QueryMethod1(nums);
         // </return_query_from_method_1>
@@ -49,10 +48,9 @@ public static class ReturnQueryFromMethod
         // </return_query_from_method_4>
 
         // <return_query_from_method_5>
-        myQuery1 =
-            from item in myQuery1
-            orderby item descending
-            select item;
+        myQuery1 = from item in myQuery1
+                   orderby item descending
+                   select item;
 
         // Execute the modified query.
         Console.WriteLine("\nResults of executing modified myQuery1:");
