@@ -7,16 +7,15 @@ public static class RuntimeFiltering
     public static void RuntimeFiltering1()
     {
         // <runtime_filtering_1>
-        int[] ids = [111, 114, 112];
+        int[] ids = [ 111, 114, 112 ];
 
-        var queryNames =
-            from student in students
-            where ids.Contains(student.ID)
-            select new
-            {
-                student.LastName,
-                student.ID
-            };
+        var queryNames = from student in students
+                         where ids.Contains(student.ID)
+                         select new
+                         {
+                             student.LastName,
+                             student.ID
+                         };
 
         foreach (var name in queryNames)
         {
@@ -30,7 +29,7 @@ public static class RuntimeFiltering
          */
 
         // Change the ids.
-        ids = [122, 117, 120, 115];
+        ids = [ 122, 117, 120, 115 ];
 
         // The query will now return different results
         foreach (var name in queryNames)
