@@ -13,7 +13,7 @@ class TestMotorcycle : Motorcycle
         var moto = new TestMotorcycle();
         moto.StartEngine();
         moto.AddGas(15);
-        int travelTime = moto.Drive(miles: 170, speed: 60);
+        int travelTime = moto.Drive(speed: 60, miles: 170);
         Console.WriteLine("Travel time: approx. {0} hours", travelTime);
     }
 }
